Update 2023.10.20_Deming_and_Simulating_Data.Rmd
The main thing I did was look for why the observed data may be different than the simulated and bootstrapped data.

I found that T231 had a lot a variation among his trials.

I also found that var is not similar between observed and simulated sets.

I tried to make the code a bit cleaner. When doing this I made the functions not work on all the males at once, so I need to continue to look into what is happening with that.