Adding data for round 4 (MDSHrna) from hard drive
This is data for the experiment I ran Spring 2023 with Misty Profitt (a former post-doc in the lab). In this experiment we had males experience a heat challenge while seeing a female. The males were dissected immediately after to preform RNA-seq on brain and syrinx.

This is various behavior data that undergrads scored and temperature logger data.