working on how to get duration for songs
Song of the timestamps recorded for files don't make sense (some are way too long and some are negative time). I am trying to get duration of the audio files themselves, but some of these (800) don't seem to have saved the cropping that Tamara did.
I am hopping to use some combination of either her written timestamps or the length of the files to determine song duration