Creating Folders and Concat Datasheet of temperature logs from camera/pi thermocouple
These loggers can be used to get an idea of what is the max temperature in the chamber. This temperature sensor is located right beside the sensor that is built into the chamber and used for setting temperature. This temperature should closely track target temperature. This sensor is ran through the pi that uploads video feed to quarrel. It logs temp even when video is not recording, it can be a potential backup for if the EasyLogger (the prefered logger) malfunctions or was forgotten.